Adapting baseapp_payments, baseapp_referrals, baseapp_pdf, baseapp_message_templates and baseapp_cloudflare_stream_field to use plugin arch
