+++ baseapp_cloudflare_stream_field/apps.py
@@ -1,7 +1,7 @@
-from django.apps import AppConfig
+from baseapp_core.plugins import BaseAppConfig
 
 
-class PackageConfig(AppConfig):
+class PackageConfig(BaseAppConfig):
     default = True
     name = "baseapp_cloudflare_stream_field"
     label = "baseapp_cloudflare_stream_field"
