+++ baseapp_message_templates/apps.py
@@ -1,7 +1,7 @@
-from django.apps import AppConfig
+from baseapp_core.plugins import BaseAppConfig
 
 
-class PackageConfig(AppConfig):
+class PackageConfig(BaseAppConfig):
     default = True
     name = "baseapp_message_templates"
     label = "baseapp_message_templates"
