+++ baseapp_payments/apps.py
@@ -1,6 +1,9 @@
-from django.apps import AppConfig
+from baseapp_core.plugins import BaseAppConfig
 
 
-class PackageConfig(AppConfig):
+class PackageConfig(BaseAppConfig):
+    default = True
     name = "baseapp_payments"
     label = "baseapp_payments"
+    verbose_name = "BaseApp Payments"
+    default_auto_field = "django.db.models.AutoField"
