+++ baseapp_payments/rest_framework/webhooks/serializers.py
@@ -1,0 +1,6 @@
+from rest_framework import serializers
+
+
+class StripeWebhookSerializer(serializers.Serializer):
+    id = serializers.CharField()
+    object = serializers.CharField()
