+++ baseapp_pdf/apps.py
@@ -1,7 +1,7 @@
-from django.apps import AppConfig
+from baseapp_core.plugins import BaseAppConfig
 
 
-class PackageConfig(AppConfig):
+class PackageConfig(BaseAppConfig):
     default = True
     name = "baseapp_pdf"
     label = "baseapp_pdf"
