+++ baseapp_referrals/admin.py
@@ -1,11 +1,10 @@
+import swapper
 from django.contrib import admin
 
-from .models import UserReferral
+UserReferral = swapper.load_model("baseapp_referrals", "UserReferral")
 
 
+@admin.register(UserReferral)
 class UserReferralAdmin(admin.ModelAdmin):
     list_display = ("id", "referrer", "referee")
     fields = ("referrer", "referee")
-
-
-admin.site.register(UserReferral, UserReferralAdmin)
