+++ pyproject.toml
@@ -42,6 +42,7 @@ dependencies = [
     "requests>=2.32.3",
     "stevedore>=5.7.0",
     "pydantic>=2.12.5",
+    "drf-spectacular==0.29.0",
 ]
 
 [project.optional-dependencies]
@@ -168,6 +169,10 @@ baseapp_pages = "baseapp_pages.plugin:PagesPlugin"
 baseapp_organizations = "baseapp_organizations.plugin:OrganizationsPlugin"
 baseapp_comments = "baseapp_comments.plugin:CommentsPlugin"
 baseapp_payments = "baseapp_payments.plugin:PaymentsPlugin"
+baseapp_referrals = "baseapp_referrals.plugin:ReferralsPlugin"
+baseapp_pdf = "baseapp_pdf.plugin:PDFPlugin"
+baseapp_message_templates = "baseapp_message_templates.plugin:MessageTemplatesPlugin"
+baseapp_claudflare_stream_field = "baseapp_cloudflare_stream_field.plugin:CloudflareStreamFieldPlugin"
 
 [tool.black]
 line-length = 100
