--- testproject/payments/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class PaymentsConfig(AppConfig):
-    name = "apps.payments"
