+++ testproject/payments/models.py
@@ -1,0 +1,15 @@
+from model_utils import FieldTracker
+
+from baseapp_payments.models import BaseCustomer, BaseSubscription
+
+
+class Customer(BaseCustomer):
+    tracker = FieldTracker(["entity"])
+
+    class Meta(BaseCustomer.Meta):
+        pass
+
+
+class Subscription(BaseSubscription):
+    class Meta(BaseSubscription.Meta):
+        pass
