+++ testproject/referrals/models.py
@@ -1,0 +1,6 @@
+from baseapp_referrals.models import BaseUserReferral
+
+
+class UserReferral(BaseUserReferral):
+    class Meta(BaseUserReferral.Meta):
+        pass
